+++ Core/Spect.Net.SpectrumEmu/Devices/Ports/SpectrumP3MemoryPagePortHandler.cs
@@ -23,17 +23,6 @@ namespace Spect.Net.SpectrumEmu.Devices.Ports
         }
 
         /// <summary>
-        /// Indicates if paging is enabled or not
-        /// </summary>
-        /// <remarks>
-        /// Port 0x7FFD, Bit 5: 
-        /// False - paging is enables
-        /// True - paging is not enabled and further output to the port
-        /// is ignored until the computer is reset
-        /// </remarks>
-        public bool PagingEnabled { get; set; }
-
-        /// <summary>
         /// Indicates the low bit of ROM selection
         /// </summary>
         /// <remarks>
@@ -101,7 +90,7 @@ namespace Spect.Net.SpectrumEmu.Devices.Ports
         public override void Reset()
         {
             base.Reset();
-            PagingEnabled = true;
+            _memoryDevice.PagingEnabled = true;
             SelectRomLow = 0;
             SelectRomHigh = 0;
         }
@@ -115,7 +104,7 @@ namespace Spect.Net.SpectrumEmu.Devices.Ports
         {
             // --- When paging is disabled, it will be enabled next time
             // --- only after reset
-            if (!PagingEnabled) return;
+            if (!_memoryDevice.PagingEnabled) return;
 
             // --- Choose the RAM bank for Slot 3 (0xc000-0xffff)
             LastSlot3Index = writeValue & 0x07;
@@ -128,7 +117,7 @@ namespace Spect.Net.SpectrumEmu.Devices.Ports
             SelectRomLow = (byte)((writeValue >> 4) & 0x01);
 
             // --- Enable/disable paging
-            PagingEnabled = (writeValue & 0x20) == 0x00;
+            _memoryDevice.PagingEnabled = (writeValue & 0x20) == 0x00;
 
             // --- Page the ROM according to current settings
             _memoryDevice.SelectRom(SelectRomHigh | SelectRomLow);