+++ VsIntegration/Spect.Net.VsPackage/ToolWindows/MemoryBankStatusUserControl.xaml.cs
@@ -1,0 +1,13 @@
+﻿namespace Spect.Net.VsPackage.ToolWindows
+{
+    /// <summary>
+    /// Interaction logic for MemoryBankStatusUserControl.xaml
+    /// </summary>
+    public partial class MemoryBankStatusUserControl
+    {
+        public MemoryBankStatusUserControl()
+        {
+            InitializeComponent();
+        }
+    }
+}