+++ CHANGELOG.md
@@ -12,3 +12,7 @@ This command works only when the Spectrum VM is paused. It sets the Program Coun
 specified _`hhhh`_ address. When the Spectum VM continues running &mdash; after the Start command &mdash;
 it will carry on from the specified address.
  
+__FEATURE__: Now, you can use the __Add VM State to project__ command in the Spectrum emulator tool window
+toolbar to save the Spectrum VM state directly into the project hierarchy.
+
+__FEATURE__: You can execute the __Load VM State file__ command from the context menu of a `.vmstate` file in Solution Explorer.